@@ -48,14 +55,12 @@
     "prop-types": "^15.6.2",
     "puppeteer": "^1.13.0",
     "raw-loader": "^1.0.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-helmet": "^5.2.0",
-    "regenerator-runtime": "^0.13.3",
     "remark": "^10.0.1",
     "remark-cli": "^7.0.0",
     "remark-frontmatter": "^1.3.2",
-    "remark-html": "^9.0.0",
     "remark-react": "^5.0.1",
     "slugg": "^1.2.1",
     "tap-min": "^2.0.0",
