@@ -1,6 +1,6 @@
 /*---
-title: Check for browser support
-description: Check for Mapbox GL browser support.
+title: Check if Mapbox GL JS is supported
+description: Check for Mapbox GL browser support. Find more information about browser support for Mapbox tools in the [Browser support](https://docs.mapbox.com/help/troubleshooting/mapbox-browser-support/) troubleshooting guide.
 tags:
   - browser-support
 pathname: /mapbox-gl-js/example/check-for-support/
