@@ -2,7 +2,7 @@
 title: Measure distances
 description: >-
   Click points on a map to create lines that measure distanced using
-  [turf.lineDistance](http://turfjs.org/static/docs/module-turf_line-distance.html).
+  [turf.length](https://turfjs.org/docs/#length).
 tags:
   - user-interaction
 pathname: /mapbox-gl-js/example/measure/
