@@ -1,6 +1,6 @@
 /*---
 title: Variable label placement
-description: Use `text-variable-anchor` to allow high priority labels to shift position to stay on the map.
+description: Use [`text-variable-anchor`](https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#layout-symbol-text-variable-anchor) to allow high priority labels to shift position to stay on the map.
 tags:
   - layers
 pathname: /mapbox-gl-js/example/variable-label-placement/
